Add auth to other instances
