+++ src/ServiceControl.Audit/App.config
@@ -25,6 +25,24 @@ These settings are only here so that we can debug ServiceControl while developin
 	<!-- options are any comma separated combination of NLog,Seq,Otlp -->
 	<add key="ServiceControl.Audit/LoggingProviders" value="NLog,Seq"/>
 	<add key="ServiceControl.Audit/SeqAddress" value="http://localhost:5341"/>
+
+    <!-- Authentication Settings (JWT with OpenID Connect) -->
+    <!-- Uncomment and configure to enable authentication -->
+    <!-- Leaving 'Authentication.Enabled' commented out defaults authentication to 'false'-->
+    <!-- <add key="ServiceControl/Authentication.Enabled" value="false" />
+    <add key="ServiceControl/Authentication.Authority" value="" />
+    <add key="ServiceControl/Authentication.Audience" value="" /> -->
+    <!-- Optional Authentication Settings (defaults shown) -->
+    <!--<add key="ServiceControl/Authentication.ValidateIssuer" value="true" />
+    <add key="ServiceControl/Authentication.ValidateAudience" value="true" />
+    <add key="ServiceControl/Authentication.ValidateLifetime" value="true" />
+    <add key="ServiceControl/Authentication.ValidateIssuerSigningKey" value="true" />
+    <add key="ServiceControl/Authentication.RequireHttpsMetadata" value="true" />-->
+    <!-- ServicePulse Authentication Settings -->
+    <!-- <add key="ServiceControl/Authentication.ServicePulse.Enabled" value="false" />
+    <add key="ServiceControl/Authentication.ServicePulse.ClientId" value="" />
+    <add key="ServiceControl/Authentication.ServicePulse.Authority" value="" />
+    <add key="ServiceControl/Authentication.ServicePulse.ApiScope" value="" /> -->
   </appSettings>
   <connectionStrings>
     <!-- DEVS - Pick a transport connection string to match chosen transport above -->
