+++ src/ServiceControl.Monitoring/Hosting/Commands/RunCommand.cs
@@ -5,6 +5,7 @@ namespace ServiceControl.Monitoring
     using Infrastructure.WebApi;
     using Microsoft.AspNetCore.Builder;
     using NServiceBus;
+    using ServiceControl.Hosting.Auth;
 
     class RunCommand : AbstractCommand
     {
@@ -13,11 +14,14 @@ namespace ServiceControl.Monitoring
             var endpointConfiguration = new EndpointConfiguration(settings.InstanceName);
 
             var hostBuilder = WebApplication.CreateBuilder();
+            hostBuilder.AddServiceControlAuthentication(settings.OpenIdConnectSettings);
             hostBuilder.AddServiceControlMonitoring((_, __) => Task.CompletedTask, settings, endpointConfiguration);
             hostBuilder.AddServiceControlMonitoringApi();
 
             var app = hostBuilder.Build();
             app.UseServiceControlMonitoring();
+            app.UseServiceControlAuthentication(authenticationEnabled: settings.OpenIdConnectSettings.Enabled);
+
             await app.RunAsync(settings.RootUrl);
         }
     }
