+++ src/ServiceControl/WebApplicationExtensions.cs
@@ -7,7 +7,7 @@ using Microsoft.AspNetCore.HttpOverrides;
 
 public static class WebApplicationExtensions
 {
-    public static void UseServiceControl(this WebApplication app, bool authenticationEnabled = false)
+    public static void UseServiceControl(this WebApplication app)
     {
         app.UseForwardedHeaders(new ForwardedHeadersOptions { ForwardedHeaders = ForwardedHeaders.All });
         app.UseResponseCompression();
@@ -15,19 +15,5 @@ public static class WebApplicationExtensions
         app.UseHttpLogging();
         app.MapHub<MessageStreamerHub>("/api/messagestream");
         app.UseCors();
-
-        // Always add middleware (harmless when not configured)
-        app.UseAuthentication();
-        app.UseAuthorization();
-
-        // Only require authorization if authentication is enabled
-        if (authenticationEnabled)
-        {
-            app.MapControllers().RequireAuthorization();
-        }
-        else
-        {
-            app.MapControllers();
-        }
     }
 }
